requirements.txt
streamlit>=1.30
tensorflow>=2.12
keras>=2.12
numpy
pillow
tqdm
h5py
gdown
